requirements.txt
pandas==1.4.3
matplotlib==3.5.2
streamlit==1.9.0
Pillow==8.2.0
neuralprophet==0.3.1